reorg types, decompose file, improve comments
