--- modules/server/src/network/common.ts
@@ -1,45 +1,0 @@
-import { ObjectValues } from '@/utils/types';
-
-/*
- * Network aggregations
- */
-export const NETWORK_AGGREGATIONS = {
-	NetworkAggregations: 'NetworkAggregations',
-	NetworkNumericAggregations: 'NetworkNumericAggregations',
-} as const;
-export type NetworkAggregation = ObjectValues<typeof NETWORK_AGGREGATIONS>;
-
-/*
- * Supported single aggregations that can be used by network aggregation
- */
-export const SUPPORTED_AGGREGATIONS = {
-	Aggregations: 'Aggregations',
-	NumericAggregations: 'NumericAggregations',
-} as const;
-
-export type SupportedAggregation = ObjectValues<typeof SUPPORTED_AGGREGATIONS>;
-
-export const SUPPORTED_AGGREGATIONS_LIST: SupportedAggregation[] = [
-	SUPPORTED_AGGREGATIONS.Aggregations,
-	SUPPORTED_AGGREGATIONS.NumericAggregations,
-];
-
-/**
- * Multiple cases when we need to map from the "single" aggregation types to the "network"
- *
- * @example
- * "Aggregations" => "NetworkAggregations"
- */
-export const singleToNetworkNameMap = new Map<SupportedAggregation, NetworkAggregation>()
-	.set(SUPPORTED_AGGREGATIONS.Aggregations, NETWORK_AGGREGATIONS.NetworkAggregations)
-	.set(SUPPORTED_AGGREGATIONS.NumericAggregations, NETWORK_AGGREGATIONS.NetworkNumericAggregations);
-
-/**
- * Multiple cases when we need to map from the "network" aggregation types to the "single"
- *
- * @example
- * "NetworkAggregations" => "Aggregations"
- */
-export const networkToSingleNameMap = new Map<NetworkAggregation, SupportedAggregation>()
-	.set(NETWORK_AGGREGATIONS.NetworkAggregations, SUPPORTED_AGGREGATIONS.Aggregations)
-	.set(NETWORK_AGGREGATIONS.NetworkNumericAggregations, SUPPORTED_AGGREGATIONS.NumericAggregations);
