+++ modules/server/src/network/resolvers/response.ts
@@ -1,5 +1,5 @@
-import { AllAggregations } from '../types/types';
-import { NetworkNode } from './networkNode';
+import { Aggregations } from '@/mapping/resolveAggregations';
+import { NetworkNode } from './aggregations';
 
 /**
  * Format response object to match GQL type defs
@@ -8,7 +8,7 @@ export const createResponse = ({
 	aggregationResults,
 	nodeInfo,
 }: {
-	aggregationResults: AllAggregations;
+	aggregationResults: Aggregations;
 	nodeInfo: NetworkNode[];
 }) => {
 	return { nodes: nodeInfo, aggregations: aggregationResults };
