+++ modules/server/src/network/typeDefs/___networkAggregations.ts
