+++ modules/server/src/network/typeDefs/index.ts
@@ -1,4 +1,4 @@
-import { SupportedNetworkFieldType } from '../types/types';
+import { SupportedNetworkFieldType } from '../setup/fields';
 import { createNetworkAggregationTypeDefs } from './aggregations';
 
 export const createTypeDefs = (networkFieldTypes: SupportedNetworkFieldType[]) => {
