+++ modules/server/src/network/types/aggregations.ts
@@ -11,7 +11,6 @@ export type Bucket = {
 };
 
 export type Aggregations = {
-	__typename: 'Aggregations';
 	bucket_count: number;
 	buckets: Bucket[];
 };
